requirements.txt
numpy
torch==1.7
matplotlib
pillow
opencv-python
pandas
jupyter
sklearn
tqdm
ipython